vsock: add net_device for packet scheduling
This patch adds a very thin net_device to the virtio-vsock and
vhost-vsock implementations.

The net_device provides the scheduling capabilities of qdisc. The user
may set qdisc policy using tc. By default, the pfifo qdisc is used. This
preserves the "configuration-free" property of vsock, and maintains
legacy behavior.

A packet is first enqueued using dev_queue_xmit(), adding the packet to
the device queue according to the qdisc. When the qdisc deems a packet
ready for dispatch, it then calls the start_xmit(). It is in
start_xmit() that packet transmission resumes as before: the packet is
added to send_pkt_list and the workqueue worker is triggered to consume
the send_pkt_list.

The result is that the packet order of send_pkt_list is determined by the
qdisc chosen by the user.

Signed-off-by: Bobby Eshleman <bobby.eshleman@bytedance.com>