vsock: add thin net_device for better packet scheduling
This patch adds a very thin net_device to the virtio-vsock and
vhost-vsock implementation.

The purpose of the net_device is simply to borrow the scheduling
capabilities of qdisc, which helps to avoid the pitfalls of
FIFO packet scheduling.

In this patch, a packet is first enqueued using dev_queue_xmit(), adding
the packet to the device queue using a qdisc. When the qdisc deems a
packet ready for dispatch, it then calls the start_xmit() function to
send the packet. It is in start_xmit() that the packet is added to
send_pkt_list and the workqueue worker is triggered to consume the
send_pkt_list (same as before).

The result is that the user can apply different scheduling policies
to vsock sockets. The default pfifo qdisc is adopted by the vsock
net_device until reconfigured by the user.

Signed-off-by: Bobby Eshleman <bobby.eshleman@bytedance.com>