vsock: create uniform initializer for virtio transport
This patch changes virtio_transport_init() to accomodate all types of
virtio transports (host, guest, and loopback) instead of just those that
use devices (host and guest).

This is done to simplify the virtio transport initialization API down to
a single function instead of requiring different initializer code based
on whether or not the transport uses a device.

Signed-off-by: Bobby Eshleman <bobby.eshleman@bytedance.com>