testing/vsock: support dgram in vsock_perf
This commit adds support for testing dgrams in vsock_perf.

It caps the buffer size at the dgram packet max (64K).

Because the datagrams are connectionless the receiver can not know to
stop receiving when the sender ends a connection (as is possible in
connection-oriented protocol). Therefore, this commit teaches the
receive-side, for dgrams only, to use the --bytes argument to control
when the receiver should stop receiving.  Once "bytes" number of bytes
have been received, it stops receiving and reports the statistics.

The usage then looks like this:

sender:

	vsock_perf --sender 2 --dgram --bytes 2G

receiver:

	vsock_perf --dgram --bytes 1G

The receiver will then report the statistics for the first 1G received.
The sender's bytes value is chosen to be bigger than the receiver's
to account for any packet loss.