testing/vsock: improve timing for rx in vsock_perf
This improves the accuracy of the time accounting on the receive side in
vsock_perf. This changes the start time sample to take place only after
traffic begins, and so removes the time between connecting but before
receiving traffic from the throughput calculation.

Additionally, this makes this function reusable by future
connection-less implementations without an out-of-band "start/stop"
signal, in which case the delay between starting the server and the
first poll() call may be very long (not just the time between a
connection and the first transmission).

Signed-off-by: Bobby Eshleman <bobby.eshleman@bytedance.com>