vsock/vmci: register h2g/g2h and dgram at same time
This patch changes vmci to register the dgram transport along with the
h2g or g2h bits. This is preliminary work to support g2h and h2g dgram.
Without the h2g or g2h bits being set, it is unclear which transport
(h2g or g2h) to choose for VSOCK_TRANSPORT_F_DGRAM.

This patch is untested.

Signed-off-by: Bobby Eshleman <bobby.eshleman@bytedance.com>