vhost/vsock: implement datagram support
This commit implements datagram support for vhost/vsock by teaching
vhost to use the common virtio transport datagram functions.

If the virtio RX buffer is too small, then the transmission is
abandoned, the packet dropped, and EHOSTUNREACH is added to the socket's
error queue.

Additionally, iov_iter_init() is now moved later (nearer to first usage)
to avoid unnecessary stores when the transmission is abandoned.

Signed-off-by: Bobby Eshleman <bobby.eshleman@bytedance.com>