vsock: add netns fallback logic and helper
vsock sockets are either in the global namespace or a local namespace.
If in the global namespace, it is accessible from any other namespace.
If in a local namespace, it is only accessible from within that
namespace. If a global namespace vsock and a local namespace vsock share
the same CID, then the global will be shadowed by the local and only the
local will be reachable.

The above mechanism is implemented using "fallback" logic where during
CID/vsock lookups any local match is used immediately, but any
global match is remembered in case there does not exist a local match.
The global match is the "fallback" in case no local match is found.

This patch adds this logic and a helper function.