vhost/vsock: use netns of process that opens the vhost-vsock-netns device
This patch assigns the network namespace of the process that opened
vhost-vsock-netns device (e.g. VMM) to the packets coming from the
guest, allowing only host sockets in the same network namespace to
communicate with the guest.

This patch also allows having different VMs, running in different
network namespace, with the same CID/port.

This patch brings namespace support only to vhost-vsock, but not
to virtio-vsock, hyperv, or vmci.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: Bobby Eshleman <bobbyeshleman@gmail.com>
---
v2 -> v3:
* use vsock_net_get_or_fallback()
* put_net() -> put_net_track
* use virtio_vsock_skb_net()/virtio_vsock_skb_set_net()
* add comment about vsock_net_has_connected()
* remove vsock->net non-null check
* rename labels out_vvm -> out_misc, out_vt -> out_core
* correct order of misc_deregister() calls in vhost_vsock_init()
* add wrappers for accesses of vsock->net, with special handling of vsock_global_net()
* use net_eq() inside vsock_net_has_connected() instead of bare ->net check
* split seqpacket_allow() stuff into new patch